[#114] Rename: rename dashboard constant file to summary
