+++ src/constants/index.ts
@@ -4,4 +4,4 @@ export * from './maxJournalCountByLevel';
 export * from './bannerData';
 export * from './roadmapData';
 export * from './plan';
-export * from './dashboard';
+export * from './summary';
